Rename gulpfile.js to gulpfile.babel.js
